fix(fuel): Don`t forget include header file

+++ FreeRTOS/Demo/Posix_GCC/data/fuel_calc.h
@@ -15,6 +15,6 @@
 
 void FuelCalcInit();
 void FuelCalcUpdate(float *distance);
-int FuelCalcAvg();
+float FuelCalcAvg();
 
 #endif // FUEL_CALC_H_
